users/views.py
from rest_framework import status
from rest_framework.response import Response
from .serializers import RegisterUserSerializer, LoginUserSerializer
from rest_framework.views import APIView
from rest_framework.authtoken.views import ObtainAuthToken
from rest_framework.authtoken.models import Token
from rest_framework import permissions


class RegisterView(APIView):
    def post(self, request):
        data = request.data
        serializer = RegisterUserSerializer(data=data)
        if serializer.is_valid(raise_exception=True):
            serializer.save()
            return Response("Successfully registered.", status=status.HTTP_201_CREATED)


class LoginView(ObtainAuthToken):
    serializer_class = LoginUserSerializer


class LogoutView(APIView):
    permission_classes = [permissions.IsAuthenticated]

    def post(self, request):
        user = request.user
        Token.objects.filter(user=user).delete()
        return Response('Successfully logged out', status=status.HTTP_200_OK)
