requirements.txt
django
Pillow
djangorestframework
django-cors-headers
drf-yasg
python-decouple==3.4
